refactor(Autoupdate/PHPVersionEnoughHigh): move to Package class
