+++ tools/autoupdate/app/Package.php
@@ -134,6 +134,65 @@ abstract class Package extends Files
         $this->extractionPath = null;
     }
 
+    /**
+     * get needed PHP version from json file from repository
+     * @return string formatted as '7.3.0', '7.3.0' is the wanted version in case of error
+     */
+    public function getNeededPHPversion(): string
+    {
+        // check format of JSON package 99.99.99
+        $matches = [];
+        if (is_string($this->minimalPhpVersion) && preg_match('/^([0-9]*)\.([0-9]*)\.([0-9]*)$/', $this->minimalPhpVersion, $matches)) {
+            return $this->minimalPhpVersion ;
+        }
+        return '7.3.0'; // just in case of error give a number
+    }
+    
+    /**
+     * get needed PHP version from json file from extracted folder
+     * @return string formatted as '7.3.0', '7.3.0' is the wanted version in case of error
+     */
+    public function getNeededPHPversionFromExtractedFolder(): string
+    {
+        $jsonPath = $this->extractionPath . 'composer.json';
+        if (file_exists($jsonPath)) {
+            $jsonFile = file_get_contents($jsonPath);
+            if (!empty($jsonFile)) {
+                $composerData = json_decode($jsonFile, true) ;
+                if (!empty($composerData['require']['php'])) {
+                    $rawNeededPHPRevision = $composerData['require']['php'];
+                    $matches = [];
+                    // accepted format '7','7.3','7.*','7.3.0','7.3.*
+                    // and these with '^', '>' or '>=' before
+                    if (preg_match('/^(\^|>=|>)?([0-9]*)(?:\.([0-9\*]*))?(?:\.([0-9\*]*))?/', $rawNeededPHPRevision, $matches)) {
+                        $major = $matches[2];
+                        $minor = $matches[3] ?? 0;
+                        $minor = ($minor == '*') ? 0 : $minor;
+                        $fix = $matches[4] ?? 0;
+                        $fix = ($fix == '*') ? 0 : $fix;
+                        return $major.'.'.$minor.'.'.$fix;
+                    }
+                }
+            }
+        }
+        return $this->getNeededPHPversion();
+    }
+    
+    /**
+     * check if current PHP version enough high
+     * @param string $neededRevision
+     * @return bool
+     */
+    public function PHPVersionEnoughHigh(?string $neededRevision = null)
+    {
+        return version_compare(
+            PHP_VERSION,
+            (empty($neededRevision))
+            ? $this->getNeededPHPversion()
+            : $neededRevision,
+            '>='
+        );
+    }
 
     /****************************************************************************
      * Méthodes privées
